+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
 
+## [1.7.0] - 2024-12-10
+- Add a new palette as easy as never. Go to settings, add a new palette name and after that add colors to your palette.
+- Also added a new palette: danger.
+- Fixed merge of palettes.
+---
+
 ## [1.6.3] - 2024-12-09
 - Add `Menu` icon. Disabled at the moment.
 - Add animations: for box, also animated border for heatmap container.